+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
-- 
+- Added emi fletching recipe category
 ### Fixed:
-- Sun and moon position
+- 
 ### Changed:
-- Updated to mc 1.20.1
+-